+++ src/main/java/com/glowingfederal/combatives/loading/CombativesCorePlugin.java
@@ -64,6 +64,7 @@ public class CombativesCorePlugin implements IFMLLoadingPlugin, IEarlyMixinLoade
                 "EntityRendererMixin",
                 "ModelBipedMixin",
                 "RenderPlayerMixin",
+                "RenderMixin",
                 "RendererLivingEntityMixin",
                 "EntityOtherPlayerMPMixin",
                 "PlayerControllerMPMixin"